daily-briefing/requirements.txt
click>=8.1,<9
python-dotenv>=1.0,<2
PyYAML>=6.0,<7
PyGithub>=2.1,<3
requests>=2.31,<3
google-api-python-client>=2.100,<3
google-auth>=2.23,<3
